finish SAD_CUDA and NCC_loss_cuda

+++ Match/cuda_localmatch.cuh
@@ -11,7 +11,7 @@ typedef unsigned char uchar;
 __global__ void addKernel(int *c, const int *a, const int *b);
 
 __global__ void SADMatch(uchar* left, uchar* right, uchar* out, int search_range, int win_size, int width, int height);
-__device__ float SADLoss(uchar* patch1, uchar* patch2, int patch_size);
+__device__ int SADLoss(uchar* patch1, uchar* patch2, int patch_size);
 
 __global__ void NCCMatch(uchar* left, uchar* right, uchar* out, int search_range, int win_size, int width, int height);
 __device__ float NCCEnergy(uchar* patch1, uchar* patch2, float left_avg, int patch_size);
